requirements.txt
# Only install AutoAWQ with x86-64
autoawq; platform_machine == "x86_64" or platform_machine == "AMD64"

# Base requirements
matplotlib
streamlit
txtai[graph,pipeline-data,pipeline-llm,similarity]
torch
Pillow
networkx
python-dotenv
